Cleaned up Student, GroupData, and Driver classes
In Driver, I broke up commandLine() into sub functions that are easier to work with. I did not change any implementation, just copied and pasted. I changed load() to return GroupData so that it's call makes it more clear as to what it does.

I removed numOfGroups() from student.

In GroupData, I added groupCount to keep track of groups and changed numOfGroups() to return this value. Insert has been changed to throw an IllegalArgumentException when the vector is not empty and there is a group length mismatch. I commented find(long id) and totals() as depreciated as a reminder to remove them after their dependencies are gone.